requirements.txt
streamlit
langchain
langchain-groq
langgraph
python-dotenv
pydantic
